requirements.txt
pandas==2.1.1
openai==0.28.1
python-dotenv==1.0.0
Flask==3.0.3
langchain==0.0.323
chromadb==0.4.15
tiktoken==0.5.1